+++ app/src/main/java/ch/epfl/cs311/wanderwave/model/remote/FirebaseConnection.kt
@@ -42,7 +42,7 @@ abstract class FirebaseConnection<T, U> {
 
   open fun addItem(item: T) {
     val itemMap = itemToMap(item)
-
+    Log.d("Conneciton","beacon map : ${itemMap}")
     db.collection(collectionName)
         .add(itemMap)
         .addOnFailureListener { e -> Log.e("Firestore", "Error adding document: ", e) }
