+++ app/src/main/java/ch/epfl/cs311/wanderwave/ui/screens/LoginScreen.kt
@@ -2,6 +2,7 @@ package ch.epfl.cs311.wanderwave.ui.screens
 
 import android.content.Context
 import android.content.ContextWrapper
+import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -14,6 +15,14 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import ch.epfl.cs311.wanderwave.model.data.Beacon
+import ch.epfl.cs311.wanderwave.model.data.Location
+import ch.epfl.cs311.wanderwave.model.data.Profile
+import ch.epfl.cs311.wanderwave.model.data.ProfileTrackAssociation
+import ch.epfl.cs311.wanderwave.model.data.Track
+import ch.epfl.cs311.wanderwave.model.remote.BeaconConnection
+import ch.epfl.cs311.wanderwave.model.remote.ProfileConnection
+import ch.epfl.cs311.wanderwave.model.remote.TrackConnection
 import ch.epfl.cs311.wanderwave.navigation.NavigationActions
 import ch.epfl.cs311.wanderwave.navigation.Route
 import ch.epfl.cs311.wanderwave.ui.components.login.LoginAppLogo
@@ -21,6 +30,9 @@ import ch.epfl.cs311.wanderwave.ui.components.login.SignInButton
 import ch.epfl.cs311.wanderwave.ui.components.login.WelcomeTitle
 import ch.epfl.cs311.wanderwave.viewmodel.LoginScreenViewModel
 import com.spotify.sdk.android.auth.AuthorizationClient
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.GlobalScope
+import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
 
 @Composable
@@ -56,10 +68,42 @@ fun LoginScreen(
     LoginAppLogo(modifier = Modifier.weight(1f))
     WelcomeTitle(modifier = Modifier.weight(4f))
     SignInButton(modifier = Modifier.weight(1f)) {
-      val intent =
-          AuthorizationClient.createLoginActivityIntent(
-              context.getActivity(), viewModel.getAuthorizationRequest())
-      launcher.launch(intent)
+      // val intent =
+      //     AuthorizationClient.createLoginActivityIntent(
+      //         context.getActivity(), viewModel.getAuthorizationRequest())
+      // launcher.launch(intent)
+      val beacon = Beacon(
+        id = "f",
+        location = Location(0.1,0.1),
+        profileAndTrack = listOf(ProfileTrackAssociation(
+          profile = Profile(
+            firstName = "New",
+            lastName = "User",
+            description = "No description",
+            numberOfLikes = 0,
+            isPublic = false,
+            spotifyUid = "newspotifyUid",
+            firebaseUid = "newfirebaseUid"),
+
+        track = Track("Sample Track ID", "Sample Track Title", "Sample Artist Name")
+
+      )))
+
+      val trackConnection = TrackConnection()
+      val profileConnection = ProfileConnection()
+
+      val beaconConnection = BeaconConnection(trackConnection = trackConnection, profileConnection = profileConnection)
+
+      beaconConnection.addItemWithId(beacon)
+
+      GlobalScope.launch {
+
+        beaconConnection.getItem("f").collect {
+          Log.d("testing", "${it}")
+        }
+      }
+
+
     }
   }
 }
